Main

+++ ProjP3.Application/DTOs/Request/ProfessorUpdateDTO.cs
@@ -8,7 +8,7 @@ namespace ProjP3.Application.DTOs.Request
 {
     public class ProfessorUpdateDTO
     {
-        public ulong IdTitulo { get; set; }
+        public int IdTitulo { get; set; }
 
         public string TxNome { get; set; } = null!;
